% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threedots.R
\name{safe_dots}
\alias{safe_dots}
\title{Only keep args a function known in \code{...}}
\usage{
safe_dots(fun)
}
\arguments{
\item{fun}{Target function to make dots dafe for.}
}
\description{
Remove all args the function \code{fun } does not know from the dots (\code{...}).
}
\examples{
library(threedots)

# toy example
foo <- \(...) {
  cat("-- START --------\n")
  print(list(...))

  cat("-- CHANGE 1: update 'a', add 'b' -------\n")
  modify_dots(a = 999, b = 2) # set a and b
  print(list(...))

  cat("-- CHANGE 2: increment 'a', remove 'b' -------\n")
  modify_dots(a = .(a) + 1, b = NULL) # set a, using a's current value, remove b
  print(list(...))
}

foo(a = 1)
}
