% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threedots-package.R
\docType{package}
\name{threedots-package}
\alias{threedots}
\alias{threedots-package}
\title{threedots: Modify the dots}
\description{
Modify the dots / ellipses (...) on the fly.
}
\author{
\strong{Maintainer}: Mark Heckmann \email{heckmann.mark@gmail.com}

}
\keyword{internal}
