% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{g}
\alias{g}
\alias{s}
\title{Modify \code{...} on the fly}
\usage{
g(x)

s(...)
}
\description{
Modify the dots (\code{...}) directly.
}
\examples{
library(threedots)

# toy example
foo <- \(...) {
  cat("-- START --------\n")
  print(list(...))

  cat("-- CHANGE 1: update 'a', add 'b' -------\n")
  s(a = 999, b = 2) # set a and b
  print(list(...))

  cat("-- CHANGE 2: increment 'a', remove 'b' -------\n")
  s(a = g(a) + 1, b = NULL) # set a, using a's current value, remove b
  print(list(...))
}

foo(a = 1)
}
